Incremental refactoring checkin.  Fixes #14
