+++ NeoAgi.AWS.CodeArtifact.Pruner/Models/EnumGeneric.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace NeoAgi
+namespace NeoAgi.AWS.CodeArtifact.Pruner.Models
 {
     /// <summary>
     /// Enum Extensions and Utilities
@@ -19,7 +19,7 @@ namespace NeoAgi
         /// <returns></returns>
         public static T ParseOrDefault(string val, T defaultVal, bool caseInsensitve = true)
         {
-            if (!Enum.TryParse<T>(val, caseInsensitve, out T parsed))
+            if (!Enum.TryParse(val, caseInsensitve, out T parsed))
             {
                 parsed = defaultVal;
             }
