+++ NeoAgi.AWS.CodeArtifact.Pruner/Program.cs
@@ -6,8 +6,8 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using NeoAgi;
 using NeoAgi.AWS.CodeArtifact.Pruner;
+using NeoAgi.AWS.CodeArtifact.Pruner.Models;
 using NeoAgi.CommandLine.Exceptions;
 using NLog;
 using NLog.Extensions.Logging;
